requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
language-tool-python==2.7.1
transformers==4.41.1
torch>=2.2.0
Pillow==10.3.0
numpy==1.26.4
